Fix tile resize logic
We do not maintain order on the tiles, so tiles appear to be randomly
placed after resize.  Taking a slightly closer look, the tiles in the
first column are the ones being placed laterally after resize (which
makes sense since onResize gets tiles in DOM order)

Fix by creating an array to keep track of tile order.  We populate this
array on init and after each page.

Also cleaned up a bit of the code:
- clean up place function a bit
- remove the Text node check out from place to fetchNextTiles since this
issue only happens on ajax
(this lets us consolidate logic of when to prevent a node from being
sent to the place function)